Move to doobie 0.5.0-M9

+++ modules/core/src/main/scala/bitpeace/FileChunk.scala
@@ -8,7 +8,7 @@ import scodec.bits.ByteVector
   */
 case class FileChunk(
   fileId: String,
-  chunkNr: Int,
+  chunkNr: Long,
   chunkData: ByteVector) {
 
   lazy val chunkLength: Long = chunkData.length.toLong
